+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.12...HEAD) - TBD
 
+### :rocket: Added
+
+- Support for validating `requests` / `httpx` / `werkzeug` responses in `Case.validate_response`. [#1718](https://github.com/schemathesis/schemathesis/issues/1718)
+
 ### :wrench: Changed
 
 - Generate at least one non-NULL character in path parameters. [#2790](https://github.com/schemathesis/schemathesis/issues/2790)
@@ -13,6 +17,8 @@
 - Do not show the "Missing test data" warning if stateful testing has non-404 responses. [#2795](https://github.com/schemathesis/schemathesis/issues/2795)
 - Restore support for validating `requests` / `httpx` / `werkzeug` responses in `validate_response` and `is_response_valid`.
 - Rename `is_response_valid` to `is_valid_response` for consistency with other functions in the codebase.
+- Change default values for `query` / `path_parameters` / `headers` / `cookies` to an empty dict instead of `None`.
+- **INTERNAL**: Add `Case.__slots__`.
 
 ### :bug: Fixed
 
@@ -24,6 +30,7 @@
 - Make header overrides case-insensitive.
 - Incorrect auth detection in `ignored_auth` when it is passed to `call` or `call_and_validate`. [#2846](https://github.com/schemathesis/schemathesis/issues/2846)
 - Incorrect check for cookie and query param based authorization in `ignored_auth`.
+- Do not include `parameters` in error message about missing HTTP method.
 
 ### :fire: Removed